refactor(agora): share survey config editor
Reuse one survey configuration editor for both conversation creation and survey editing so the two flows keep behavior in sync.

The shared editor now owns survey question editing, template insertion, validation highlighting, remove confirmations, and edit-only semantic invalidation controls via optional props. It also hides the optional/required survey toggle until at least one question exists and creates the first question/template in a single model update to avoid requiring a second click.

Deploy: agora
